Experiment with some additional example domains
Curiously, the plan produced for problem #1 of the room domain is not
complete. Based on the problem definition, John should close the gate
behind him after he goes into the room. But for some reason, and no
matter what I do, that last action of closing the gate is never called.